requirements.txt
wget==3.2
Flask==1.1.1
requests==2.22.0
python-docx==0.8.10